+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 🐛 Bug Fix
 
-- Fix for when the app would sometimes crash checking the older generation games
+- Fix for when the app would sometimes crash checking the coverage in Team Builder for older generation games
 - Added Spiritomb to Route 209 encounters
 - Added encounters to Grand Underground